+++ Source/HLEGraphics/ConvertImage.cpp
@@ -482,8 +482,8 @@ static void ConvertCI4_Row_To_8888( NativePf8888 * dst, const u8 * src, u32 src_
 	{
 		u8 b {src[src_offset ^ F]};
 
-		u32 bhi {(b&0xf0)>>4};
-		u32 blo {(b&0x0f)};
+		u32 bhi {(u32)(b&0xf0)>>4};
+		u32 blo {(u32)(b&0x0f)};
 
 		dst[ x + 0 ] = palette[ bhi ];	// Remember palette has already been swapped
 		dst[ x + 1 ] = palette[ blo ];
@@ -496,7 +496,7 @@ static void ConvertCI4_Row_To_8888( NativePf8888 * dst, const u8 * src, u32 src_
 	{
 		u8 b {src[src_offset ^ F]};
 
-		u8 bhi {(b&0xf0)>>4};
+		u8 bhi {(u8)((b&0xf0)>>4)};
 
 		dst[width-1] = palette[ bhi ];	// Remember palette has already been swapped
 	}
